Fix bugs, small refactors
